requirements.txt
asgiref==3.8.1
Django==5.0.3
django-phonenumber-field==7.3.0
mysqlclient==2.2.4
phonenumberslite==8.13.33
pkgconfig==1.5.5
sqlparse==0.4.4
